demo/requirements.txt
streamlit>=1.38.0
websocket-client>=1.8.0
boto3>=1.34.0
